Rename to object_store_registry_impl.rs
