+++ src/infra/core/src/repos/mod.rs
@@ -29,7 +29,7 @@ mod object_repository_local_fs;
 mod object_repository_s3;
 mod object_store_builder_local_fs;
 mod object_store_builder_s3;
-mod object_store_registy_impl;
+mod object_store_registry_impl;
 mod reference_repository_impl;
 
 pub use dataset_factory_impl::*;
@@ -53,5 +53,5 @@ pub use object_repository_local_fs::*;
 pub use object_repository_s3::*;
 pub use object_store_builder_local_fs::*;
 pub use object_store_builder_s3::*;
-pub use object_store_registy_impl::*;
+pub use object_store_registry_impl::*;
 pub use reference_repository_impl::*;
