+++ src/infra/core/src/repos/object_store_registry_impl.rs
